requirements.txt
# Core scientific computing
numpy>=1.24.0,<2.0.0
scipy>=1.10.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
pandas>=2.0.0,<3.0.0

# Deep learning
torch>=2.0.0,<2.6.0
botorch>=0.9.0,<0.13.0
torchmetrics>=1.0.0,<2.0.0
sentence-transformers>=3.0.0,<4.0.0

# Visualization
matplotlib>=3.7.0,<4.0.0
tueplots>=0.0.14,<0.1.0
seaborn>=0.12.0,<1.0.0
umap-learn>=0.5.6,<1.0.0

# Jupyter support
jupyter>=1.0.0,<2.0.0
jupyter-cache>=0.6.0,<1.0.0

# Utilities
pyyaml>=6.0.0,<7.0.0
tqdm>=4.65.0,<5.0.0
huggingface_hub>=0.16.0,<1.0.0
edge-tts>=6.1.0,<7.0.0
